requirements.txt
Django==2.0.8
django-cors-headers==2.4.0
django-mssql==1.8
django-pyodbc-azure==2.0.8.0
djangorestframework==3.8.2
mysqlclient==1.3.13
pyodbc==4.0.24
pytz==2018.5
wfastcgi==3.0.0
